+++ R/geom_boxjitter.R
@@ -9,11 +9,9 @@
 #' @param jitter.colour,jitter.color,jitter.fill,jitter.shape,jitter.size,jitter.stroke,jitter.alpha
 #' Default aesthetics for jitter, set to `NULL` to inherit from the aesthetics used for the box.
 #' 
-#' @param jitter.width Width passed to position_jitter. Defaults to half the width of the boxplot.
+#' @param jitter.position Position object used for calculating jitter (defaults to `ggplot2::PositionJitter`).
 #' 
-#' @param jitter.height Height passed to position_jitter. Defaults to 40 percent of the resolution.
-#' 
-#' @param jitter.seed Seed passed to position_jitter for reproducible jittering.
+#' @param jitter.params Parameters passed to `jitter.position` (for `ggplot2::PositionJitter`, this is `width`, `height` and `seed`).
 #' 
 #' @param boxplot.expand Defaults to `FALSE`. If set to `TRUE`, the full boxplots will be plotted.
 #' 
@@ -31,8 +29,7 @@
 #'                  genotype = factor(sample(1:3, 150, replace = TRUE)))
 #' 
 #' ggplot(df) + geom_boxjitter(aes(x = gender, y = score, fill = genotype),
-#'                             jitter.shape = 21, jitter.color = NA, 
-#'                             jitter.height = 0, jitter.width = 0.04,
+#'                             jitter.shape = 21, jitter.color = NA,
 #'                             outlier.color = NA, errorbar.draw = TRUE) +
 #'   scale_fill_manual(values = c("#CF3721", "#31A9B8", "#258039")) +
 #'   theme_minimal()
@@ -54,9 +51,11 @@ geom_boxjitter <- function(mapping = NULL, data = NULL,
                            jitter.size = 1.5,
                            jitter.stroke = 0.5,
                            jitter.alpha = NULL,
-                           jitter.width = NULL,
-                           jitter.height = NULL,
-                           jitter.seed = NULL,
+                           # jitter.width = NULL,
+                           # jitter.height = NULL,
+                           # jitter.seed = NULL,
+                           jitter.position = ggplot2::PositionJitter, # TODO import,
+                           jitter.params = list("width" = NULL, "height" = NULL),
                            boxplot.expand = FALSE,
                            notch = FALSE,
                            notchwidth = 0.5,
@@ -98,9 +97,11 @@ geom_boxjitter <- function(mapping = NULL, data = NULL,
       jitter.size = jitter.size,
       jitter.stroke = jitter.stroke,
       jitter.alpha = jitter.alpha,
-      jitter.width = jitter.width,
-      jitter.height = jitter.height,
-      jitter.seed = jitter.seed,
+      # jitter.width = jitter.width,
+      # jitter.height = jitter.height,
+      # jitter.seed = jitter.seed,
+      jitter.position = jitter.position,
+      jitter.params = jitter.params,
       boxplot.expand = boxplot.expand,
       notch = notch,
       notchwidth = notchwidth,
@@ -119,7 +120,7 @@ geom_boxjitter <- function(mapping = NULL, data = NULL,
 #' @importFrom ggplot2 alpha ggproto GeomBoxplot aes GeomSegment GeomPoint GeomCrossbar resolution PositionJitter
 #' @importFrom grid grobTree
 #' @export
-GeomBoxJitter <- ggproto("GeomBoxJitter", GeomBoxplot,
+GeomBoxJitter <- ggproto("GeomBoxJitter2", GeomBoxplot,
   default_aes = aes(weight = 1, colour = "grey20", fill = "white", size = 0.5,
                     alpha = NA, shape = 19, linetype = "solid"),
   
@@ -134,12 +135,14 @@ GeomBoxJitter <- ggproto("GeomBoxJitter", GeomBoxplot,
                         jitter.colour = NULL, jitter.fill = NULL,
                         jitter.shape = 19, jitter.size = 1.5,
                         jitter.stroke = 0.5, jitter.alpha = NULL,
-                        jitter.width = NULL, jitter.height = NULL,
-                        jitter.seed = NULL,
+                        # jitter.width = NULL, jitter.height = NULL,
+                        # jitter.seed = NULL, 
+                        jitter.position = ggplot2::PositionJitter,
+                        jitter.params = list("width" = NULL, "height" = NULL),
                         boxplot.expand = FALSE,
                         notch = FALSE, notchwidth = 0.5, varwidth = FALSE,
                         errorbar.draw = FALSE, errorbar.length = 0.5) {
-
+    
     xrange <- data$xmax - data$xmin
     
     common <- data.frame(
@@ -203,23 +206,40 @@ GeomBoxJitter <- ggproto("GeomBoxJitter", GeomBoxplot,
     )
     
     if (!boxplot.expand) {
-      jitter.width <- jitter.width %||% 
-        (data$xmax - (box$xmax + (data$xmax - box$xmax) / 1.75))
-      jitter.height <- jitter.height %||% (
-        resolution(data$jitter_y[[1]], zero = FALSE) * 0.4)
+      # jitter.width <- jitter.width %||% 
+      #   (data$xmax - (box$xmax + (data$xmax - box$xmax) / 1.75))
+      # jitter.height <- jitter.height %||% (
+      #   resolution(data$jitter_y[[1]], zero = FALSE) * 0.4)
+      # jitter_df <- data.frame(
+      #   width = jitter.width,
+      #   height = jitter.height
+      #   )
+      # if (!is.null(jitter.seed)) jitter_df$seed = jitter.seed
       
-      jitter_df <- data.frame(
-        width = jitter.width,
-        height = jitter.height
-        )
-      
-      if (!is.null(jitter.seed)) jitt_df$seed = jitter.seed
-      
-      jitter_positions <- PositionJitter$compute_layer(
-        data.frame(x = box$xmax + (data$xmax - box$xmax) / 2,
-                   y = data$jitter_y[[1]]),
-        jitter_df
-      )
+      jitter_position_data <- data.frame(x = box$xmax + (data$xmax - box$xmax) / 2,
+                                         y = data$jitter_y[[1]], PANEL = 1, group = -1)
+      if (is(jitter.position, "PositionJitter")) {
+        jitter.params$width <- jitter.params$width %||% 
+          (data$xmax - (box$xmax + (data$xmax - box$xmax) / 1.75))
+        jitter.params$height <- jitter.params$height %||% 
+          (resolution(data$jitter_y[[1]], zero = FALSE) * 0.4)
+
+        jitter_positions <- jitter.position$compute_layer(jitter_position_data,
+                                                          jitter.params) # TODO compute_panel for PositionJitter.
+      } else {
+        jitter_positions <- jitter.position$compute_panel(jitter_position_data, 
+                                                          jitter.params) 
+        # Scale back to original range
+        jitter_positions$x <- ((data$xmax - 0.025) - box$xmax - 0.025) * (jitter_positions$x - 
+          min(jitter_positions$x)) / (max(jitter_positions$x) - min(jitter_positions$x)) + box$xmax + 0.025
+        
+      }
+
+      # jitter_positions <- PositionJitter$compute_layer(
+      #   data.frame(x = box$xmax + (data$xmax - box$xmax) / 2,
+      #              y = data$jitter_y[[1]]),
+      #   jitter_df
+      # )
   
       if (outlier.intersect & length(data$outliers[[1]]) >= 1) {
         outlier_inds <- which(data$jitter_y[[1]] %in% data$outliers[[1]])
@@ -235,17 +255,17 @@ GeomBoxJitter <- ggproto("GeomBoxJitter", GeomBoxplot,
         shape = jitter.shape %||% data$shape[1],
         size = jitter.size %||% data$size[1],
         stroke = jitter.stroke %||% data$stroke[1],
-        fill = NA,
         alpha = jitter.alpha %||% data$alpha[1],
         stringsAsFactors = FALSE
       )
+      
       
       jitter_grob <- GeomPoint$draw_panel(jitt, panel_params, coord)
     } else {
       jitter_grob <- NULL
     }
     
-    if (!is.null(data$outliers) && length(data$outliers[[1]] >= 1)) {
+    if (!is.null(data$outliers) && length(data$outliers[[1]] >= 1)) { # TODO make outlier plotting optional
       outliers <- data.frame(
         y = if (outlier.intersect & !boxplot.expand) jitter_outlier$y else 
           data$outliers[[1]],
@@ -256,7 +276,6 @@ GeomBoxJitter <- ggproto("GeomBoxJitter", GeomBoxplot,
         shape = outlier.shape %||% data$shape[1],
         size = outlier.size %||% data$size[1],
         stroke = outlier.stroke %||% data$stroke[1],
-        fill = NA,
         alpha = outlier.alpha %||% data$alpha[1],
         stringsAsFactors = FALSE
       )